NBLAST: use Postgres' Large Objects for similarity score storage
This migration also adds three new parameter fields to NBLAST
similarities (min length, min soma length, soma tags) and changes the
data type of the result scores. Instead of using a 2D float array
(real[][]), a Postgres large object is used, which is only referenced as
an OID and is stored separately. This makes access slightly more
complex, but is needed in order to store results larger than 2GB. Those
large objects are deleted through tigger functions if the referencing
nblast_similarity entry is deleted (or the table is truncated).

This is done in multiple steps: first the schema changes are made, but
the new scoring column is added with a temporary suffix. Then the
existing scoring is translated and finally the old scoring is deleted
and the new scoring column renamed.

After this change, there will also be no more change history for the
nblast_similarity table. This isn't really needed for this set of
results and would complicate things slightly for this change.

Additionally, the table nblast_similarity_score is added, which will
allow the relational storage of query results. This should be fine in
most cases, because the vast majority of skeletons is not similar (in
the NBLAST sense) to each other.